requirements.txt
fastapi>=0.102.0,<0.110.0
uvicorn[standard]==0.22.0
python-multipart==0.0.6
google-generativeai>=0.8.0,<0.9.0
SQLAlchemy>=2.0.0,<2.1.0
pydantic>=2.0.0,<2.2.0
python-dotenv==1.0.1
Jinja2==3.1.4